galaxycache.go, hotcache.go: move hc types and methods to new file
